# contracts/openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.UpdateUserRolesRequest:
    properties:
      roles:
        description: |-
          Roles is the list of roles to assign to the user.
          Valid roles: "user", "admin", "blocked"
          Example: ["user", "admin"]
        example:
        - user
        - admin
        items:
          type: string
        type: array
    type: object
  dto.UpdateUserRolesResponse:
    properties:
      email:
        description: |-
          Email is the user's email.
          Example: "alice@example.com"
        example: alice@example.com
        type: string
      roles:
        description: |-
          Roles is the updated list of roles.
          Example: ["user", "admin"]
        example:
        - user
        - admin
        items:
          type: string
        type: array
      updated_at:
        description: |-
          UpdatedAt is the timestamp when the user was updated.
          Example: "2025-09-14T22:01:02Z"
        example: "2025-09-14T22:01:02Z"
        format: date-time
        type: string
      user_id:
        description: |-
          ID is the user's unique identifier.
          Example: 42
        example: 42
        type: integer
      username:
        description: |-
          Username is the user's username.
          Example: "alice"
        example: alice
        type: string
    type: object
...
info:
  contact:
    email: felipe.lechitz@gmail.com
    name: Lechitz
    url: https://github.com/lechitz
  description: Public and internal REST API for AionAPI. Swagger (OpenAPI 3.x) is
    generated via swaggo.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lechitz/AionApi
  title: AionAPI — REST API Documentation
  version: 0.1.0
paths:
...
  /admin/users/{user_id}/roles:
    put:
      consumes:
      - application/json
      deprecated: true
      description: |-
        Generic role update endpoint with minimal hierarchy validation.
        **DEPRECATED**: Prefer specific endpoints for better validation:
        - PUT /admin/users/{id}/promote-admin (adds admin role)
        - PUT /admin/users/{id}/demote-admin (removes admin role)
        - PUT /admin/users/{id}/block (blocks user)
        - PUT /admin/users/{id}/unblock (unblocks user)
        This endpoint provides maximum flexibility but less validation.
        Use only for special administrative operations or bulk scripts.
      parameters:
      - description: User ID (uint64)
        in: path
        name: user_id
        required: true
        type: integer
      - description: New roles to assign
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Roles updated
          schema:
            $ref: '#/definitions/dto.UpdateUserRolesResponse'
        "400":
          description: Invalid request or user_id
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden - not an admin
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Update user roles (Legacy)
      tags:
      - Admin
...
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the current authenticated session (token or cookie)
        and clears the auth cookie.
      produces:
      - application/json
      responses:
        "204":
          description: Logout succeeded (no content)
          headers:
            Set-Cookie:
              description: auth_token=deleted; Path=/; Max-Age=0; HttpOnly; Secure
                (if enabled)
              type: string
          schema:
            type: string
        "401":
          description: Unauthorized or missing user context
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Logout current user
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Use "Bearer {token}" (JWT).
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: Session-based authentication cookie (e.g., Redis-backed).
    in: cookie
    name: auth_token
    type: apiKey
    x-graphqlplayground: /graphql/ // Cross-reference to GraphQL Playground (not covered
      by Swagger)
swagger: "2.0"